@@ -1,13 +1,10 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const bedrock = require('bedrock');
-const {config} = bedrock;
-require('bedrock-validation');
-require('bedrock-app-identity');
-require('bedrock-meter-usage-reporter');
+import {config} from 'bedrock';
+import 'bedrock-validation';
+import 'bedrock-app-identity';
+import 'bedrock-meter-usage-reporter';
 
 const namespace = 'kms-http';
 const cfg = config[namespace] = {};
