@@ -1,28 +1,27 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const {asyncHandler} = require('bedrock-express');
-const bedrock = require('bedrock');
-const brZCapStorage = require('bedrock-zcap-storage');
-const {config, util: {BedrockError}} = bedrock;
-const cors = require('cors');
-const {
-  defaultModuleManager: moduleManager,
+import * as bedrock from 'bedrock';
+import * as brZCapStorage from 'bedrock-zcap-storage';
+import * as helpers from './helpers.js';
+import * as middleware from './middleware.js';
+import {asyncHandler} from 'bedrock-express';
+import cors from 'cors';
+import {createValidateMiddleware as validate} from 'bedrock-validation';
+import {
+  defaultModuleManager as moduleManager,
   keystores
-} = require('bedrock-kms');
-const {generateRandom} = require('@digitalbazaar/webkms-switch');
-const helpers = require('./helpers');
-const {meters} = require('bedrock-meter-usage-reporter');
-const middleware = require('./middleware');
-const {
+} from 'bedrock-kms';
+import {generateRandom} from '@digitalbazaar/webkms-switch';
+import {meters} from 'bedrock-meter-usage-reporter';
+import {
   postKeystoreBody,
   updateKeystoreConfigBody,
   postRevocationBody
-} = require('../schemas/bedrock-kms-http');
-const {reportOperationUsage, SERVICE_TYPE} = require('./metering');
-const {validator: validate} = require('./validator');
+} from '../schemas/bedrock-kms-http.js';
+import {reportOperationUsage, SERVICE_TYPE} from './metering.js';
+
+const {config, util: {BedrockError}} = bedrock;
 
 bedrock.events.on('bedrock-express.configure.routes', app => {
   const cfg = config['kms-http'];
