@@ -1,8 +1,5 @@
 /*!
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const bedrock = require('bedrock');
-
-module.exports = bedrock.loggers.get('app').child('bedrock-kms-http');
+import {loggers} from 'bedrock';
+export const logger = loggers.get('app').child('bedrock-kms-http');
