@@ -1,23 +1,20 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
-
-const bedrock = require('bedrock');
-const {
-  defaultModuleManager: moduleManager,
+import * as bedrock from 'bedrock';
+import * as brZCapStorage from 'bedrock-zcap-storage';
+import {
+  defaultModuleManager as moduleManager,
   keystores
-} = require('bedrock-kms');
-const brZCapStorage = require('bedrock-zcap-storage');
-const {meters} = require('bedrock-meter-usage-reporter');
-const logger = require('./logger');
+} from 'bedrock-kms';
+import {meters} from 'bedrock-meter-usage-reporter';
+import {logger} from './logger.js';
 
 // configure usage aggregator for webkms meters
-const SERVICE_TYPE = 'webkms';
-exports.SERVICE_TYPE = SERVICE_TYPE;
+export const SERVICE_TYPE = 'webkms';
 meters.setAggregator({serviceType: SERVICE_TYPE, handler: _aggregateUsage});
 
-exports.reportOperationUsage = function({req}) {
+export function reportOperationUsage({req}) {
   // do not wait for usage to be reported
   const {keystore, keystore: {meterId: id}} = req.webkms;
   meters.use({id, operations: 1}).catch(
