@@ -1,19 +1,17 @@
 /*
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import * as bedrock from 'bedrock';
+import * as helpers from './helpers.js';
+import {KeystoreConfigStorage} from '@digitalbazaar/webkms-switch';
+import {keystores} from 'bedrock-kms';
 
-const {KeystoreConfigStorage} = require('@digitalbazaar/webkms-switch');
-const {keystores} = require('bedrock-kms');
-const bedrock = require('bedrock');
 const {util: {BedrockError}} = bedrock;
-const helpers = require('./helpers');
 
 // load config defaults
-require('./config');
+import './config.js';
 
-module.exports = class BedrockKeystoreConfigStorage
-  extends KeystoreConfigStorage {
+export class BedrockKeystoreConfigStorage extends KeystoreConfigStorage {
   async get({id, req} = {}) {
     const {config: keystoreConfig} = await keystores.get({id});
 
