@@ -1,19 +1,17 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import forwarded from 'forwarded';
+import {Netmask} from 'netmask';
 
-const forwarded = require('forwarded');
-const {Netmask} = require('netmask');
-
-exports.getKeystoreId = ({host, req, localId, routes}) => {
+export function getKeystoreId({host, req, localId, routes}) {
   if(!host) {
     host = req.get('host');
   }
   return `https://${host}${routes.keystores}/${localId}`;
-};
+}
 
-exports.verifyRequestIp = ({keystoreConfig, req}) => {
+export function verifyRequestIp({keystoreConfig, req}) {
   const {ipAllowList} = keystoreConfig;
   if(!ipAllowList) {
     return {verified: true};
@@ -35,4 +33,4 @@ exports.verifyRequestIp = ({keystoreConfig, req}) => {
   }
 
   return {verified: false};
-};
+}
