@@ -1,11 +1,11 @@
 /*!
- * Copyright (c) 2019-2021 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2021-2022 Digital Bazaar, Inc. All rights reserved.
  */
 'use strict';
 
-require('bedrock-express');
+// translate `main.js` to CommonJS
+require = require('esm')(module);
+module.exports = require('./main.js');
 
-require('./http');
-
-// load config defaults
-require('./config');
+// backwards compatibility: export full API as `default`
+module.exports.default = module.exports;
