@@ -1,27 +1,27 @@
 /*!
  * Copyright (c) 2019-2022 Digital Bazaar, Inc. All rights reserved.
  */
-'use strict';
+import * as bedrock from 'bedrock';
+import * as brZCapStorage from 'bedrock-zcap-storage';
+import * as helpers from './helpers.js';
+import {asyncHandler} from 'bedrock-express';
+import {
+  authorizeZcapInvocation as _authorizeZcapInvocation,
+  authorizeZcapRevocation as _authorizeZcapRevocation
+} from '@digitalbazaar/ezcap-express';
+import {BedrockKeystoreConfigStorage} from './BedrockKeystoreConfigStorage.js';
+import {createOperationMiddleware} from '@digitalbazaar/webkms-switch';
+import {CryptoLD} from 'crypto-ld';
+import {
+  defaultDocumentLoader as documentLoader,
+  defaultModuleManager as moduleManager
+} from 'bedrock-kms';
+import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
+import {Ed25519VerificationKey2020} from
+  '@digitalbazaar/ed25519-verification-key-2020';
+import {reportOperationUsage} from './metering.js';
 
-const {asyncHandler} = require('bedrock-express');
-const {
-  authorizeZcapInvocation, authorizeZcapRevocation
-} = require('@digitalbazaar/ezcap-express');
-const bedrock = require('bedrock');
-const BedrockKeystoreConfigStorage = require('./BedrockKeystoreConfigStorage');
-const brZCapStorage = require('bedrock-zcap-storage');
-const {createOperationMiddleware} = require('@digitalbazaar/webkms-switch');
 const {config, util: {BedrockError}} = bedrock;
-const {CryptoLD} = require('crypto-ld');
-const {
-  defaultDocumentLoader: documentLoader,
-  defaultModuleManager: moduleManager
-} = require('bedrock-kms');
-const {Ed25519Signature2020} = require('@digitalbazaar/ed25519-signature-2020');
-const {Ed25519VerificationKey2020} = require(
-  '@digitalbazaar/ed25519-verification-key-2020');
-const helpers = require('./helpers');
-const {reportOperationUsage} = require('./metering');
 
 const storage = new BedrockKeystoreConfigStorage();
 
@@ -31,7 +31,7 @@ cryptoLd.use(Ed25519VerificationKey2020);
 
 // creates middleware to get the keystore config for the current request and
 // caches it in `req.webkms.keystore`
-exports.createGetKeystoreConfig = function({routes}) {
+export function createGetKeystoreConfig({routes}) {
   return asyncHandler(async function _getKeystoreConfig(req, res, next) {
     if(!req.webkms) {
       const keystoreId = helpers.getKeystoreId(
@@ -41,10 +41,10 @@ exports.createGetKeystoreConfig = function({routes}) {
     }
     next();
   });
-};
+}
 
 // creates middleware for handling KMS operations
-exports.createKmsOperationMiddleware = function() {
+export function createKmsOperationMiddleware() {
   const {kmsOperationOptions} = config['kms-http'];
   return createOperationMiddleware({
     ...kmsOperationOptions,
@@ -53,25 +53,25 @@ exports.createKmsOperationMiddleware = function() {
     inspectCapabilityChain,
     onError, onSuccess: reportOperationUsage, suiteFactory
   });
-};
+}
 
 // calls ezcap-express's authorizeZcapInvocation w/constant params, exposing
 // only those params that change in this module
-exports.authorizeZcapInvocation = function({
+export function authorizeZcapInvocation({
   getExpectedValues, getRootController
 }) {
-  return authorizeZcapInvocation({
+  return _authorizeZcapInvocation({
     documentLoader, getExpectedValues, getRootController,
     getVerifier,
     inspectCapabilityChain,
     onError,
     suiteFactory
   });
-};
+}
 
 // creates middleware for keystore route authz checks
-exports.authorizeKeystoreZcapInvocation = function() {
-  return exports.authorizeZcapInvocation({
+export function authorizeKeystoreZcapInvocation() {
+  return authorizeZcapInvocation({
     async getExpectedValues({req}) {
       return {
         host: config.server.host,
@@ -83,11 +83,11 @@ exports.authorizeKeystoreZcapInvocation = function() {
       return req.webkms.keystore.controller;
     }
   });
-};
+}
 
 // creates middleware for revocation of zcaps for keystores
-exports.authorizeZcapRevocation = function() {
-  return authorizeZcapRevocation({
+export function authorizeZcapRevocation() {
+  return _authorizeZcapRevocation({
     documentLoader,
     expectedHost: config.server.host,
     async getRootController({req}) {
@@ -99,7 +99,7 @@ exports.authorizeZcapRevocation = function() {
     onError,
     suiteFactory
   });
-};
+}
 
 // hook used to verify zcap invocation HTTP signatures
 async function getVerifier({keyId, documentLoader}) {
